Twitter Parser Prototype
The dataset Malaysian Airlines is the crawled tweets. It consists of
tweets, retweets and replies

The Parser will parse these three and store it into different csv’s. At
present the retweets parsing is completed, for the rest some fine
tuning is required.

Feel free to make changes!

Hope this gets you started!